feat: add auth middleware, user profile, edit profile, edit photo, edit password

+++ repository/Repository.go
@@ -7,7 +7,7 @@ import (
 
 type UserRepository interface {
 	GetUser() ([]entity.User, error)
-	GetUserByID(id int64) (entity.User, error)
+	GetUserByID(id int) (entity.User, error)
 	GetUserByUsername(username string) (entity.User, error)
 	GetUserByEmail(email string) (entity.User, error)
 	CreateUser(user payloads.CreateRequest) error
